backend/requirements.txt
gunicorn
fastapi
uvicorn
python-dotenv
requests
pydantic
python-multipart
Pillow
aiofiles
openai
pydub
gtts
cv2
moviepy
ruff